requirements.txt
boto3
flask
python-decouple
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2
MarkupSafe==1.1.1
Werkzeug
